Update ServerProgram.py
Fixed navigating back to the previous subfolder (subfolder to main folder); made both the current_dir and BASE_DIR absolute paths so that common path works. Navigating back will now display the entire folder path as well, which is a little janky but it works. Also fixed an issue where downloading a file would sometimes result in 0.00000 seconds and division by zero would occur, resulting in a crash. Added a '+ 0.000001' to the 'download_time' to prevent the issue from occurring, similar to the uploading fix.